resolved problems with pow() on windows, deleted yapps.exe

+++ environment/ordinates.cpp
@@ -221,7 +221,7 @@ double Vec3::LocalLength(void) const {
 }
 
 double Vec3::GlobalLength(void) const {
-    return sqrt(pow(_global.x * GRID_UNIT + x, 2.0) + pow(_global.y * GRID_UNIT + y, 2.0) + pow(_global.z * GRID_UNIT + z, 2.0));
+    return sqrt(pow((double)_global.x * GRID_UNIT + x, (double)2.0) + pow((double)_global.y * GRID_UNIT + y, (double)2.0) + pow((double)_global.z * GRID_UNIT + z, (double)2.0));
 }
 
 double Vec3::AngleToGlobal(const Vec3& vgl) const {
